+++ src/CStorage.sol
@@ -9,12 +9,11 @@ contract CStorage {
     address public borrowable0;
     address public borrowable1;
     address public simpleUniswapOracle;
-    uint public safetyMarginSqrt = 1.58113883e18; //safetyMargin: 250%
-    uint public liquidationIncentive = 1.02e18; //2%
-    uint public liquidationFee = 0.02e18; //2%
+    uint256 public safetyMarginSqrt = 1.58113883e18; //safetyMargin: 250%
+    uint256 public liquidationIncentive = 1.02e18; //2%
+    uint256 public liquidationFee = 0.02e18; //2%
 
     function liquidationPenalty() public view returns (uint) {
         return liquidationIncentive + liquidationFee;
     }
 }
-
