+++ src/ImpermaxERC20.sol
@@ -8,20 +8,24 @@ contract ImpermaxERC20 {
     string public name;
     string public symbol;
     uint8 public decimals = 18;
-    uint public totalSupply;
+    uint256 public totalSupply;
     mapping(address => uint) public balanceOf;
     mapping(address => mapping(address => uint)) public allowance;
 
     bytes32 public DOMAIN_SEPARATOR;
     mapping(address => uint) public nonces;
 
-    event Transfer(address indexed from, address indexed to, uint value);
-    event Approval(address indexed owner, address indexed spender, uint value);
+    event Transfer(address indexed from, address indexed to, uint256 value);
+    event Approval(
+        address indexed owner,
+        address indexed spender,
+        uint256 value
+    );
 
     function _setName(string memory _name, string memory _symbol) internal {
         name = _name;
         symbol = _symbol;
-        uint chainId;
+        uint256 chainId;
         assembly {
             chainId := chainid()
         }
@@ -38,35 +42,38 @@ contract ImpermaxERC20 {
         );
     }
 
-    function _mint(address to, uint value) internal {
+    function _mint(address to, uint256 value) internal {
         totalSupply = (totalSupply + value);
         balanceOf[to] = (balanceOf[to] + value);
         emit Transfer(address(0), to, value);
     }
 
-    function _burn(address from, uint value) internal {
+    function _burn(address from, uint256 value) internal {
         balanceOf[from] = (balanceOf[from] - value);
         totalSupply = (totalSupply - value);
         emit Transfer(from, address(0), value);
     }
 
-    function _approve(address owner, address spender, uint value) private {
+    function _approve(address owner, address spender, uint256 value) private {
         allowance[owner][spender] = value;
         emit Approval(owner, spender, value);
     }
 
-    function _transfer(address from, address to, uint value) internal {
+    function _transfer(address from, address to, uint256 value) internal {
         balanceOf[from] = (balanceOf[from] - value);
         balanceOf[to] = (balanceOf[to] + value);
         emit Transfer(from, to, value);
     }
 
-    function approve(address spender, uint value) external returns (bool) {
+    function approve(address spender, uint256 value) external returns (bool) {
         _approve(msg.sender, spender, value);
         return true;
     }
 
-    function transfer(address to, uint value) external override returns (bool) {
+    function transfer(
+        address to,
+        uint256 value
+    ) external override returns (bool) {
         _transfer(msg.sender, to, value);
         return true;
     }
@@ -74,7 +81,7 @@ contract ImpermaxERC20 {
     function transferFrom(
         address from,
         address to,
-        uint value
+        uint256 value
     ) external returns (bool) {
         if (allowance[from][msg.sender] != type(uint256).max) {
             allowance[from][msg.sender] = (allowance[from][msg.sender] - value);
@@ -86,8 +93,8 @@ contract ImpermaxERC20 {
     function _checkSignature(
         address owner,
         address spender,
-        uint value,
-        uint deadline,
+        uint256 value,
+        uint256 deadline,
         uint8 v,
         bytes32 r,
         bytes32 s,
@@ -124,8 +131,8 @@ contract ImpermaxERC20 {
     function permit(
         address owner,
         address spender,
-        uint value,
-        uint deadline,
+        uint256 value,
+        uint256 deadline,
         uint8 v,
         bytes32 r,
         bytes32 s
